Rework Layout to mimic Wikidata Query Service
