+++ package.json
@@ -24,8 +24,6 @@
     "@codingame/monaco-vscode-theme-defaults-default-extension": "^5.0.1",
     "@codingame/monaco-vscode-theme-service-override": "^5.0.1",
     "@reduxjs/toolkit": "^1.9.7",
-    "@types/react-window": "^1.8.7",
-    "@types/react-window-infinite-loader": "^1.0.8",
     "@types/wicg-file-system-access": "^2023.10.2",
     "bootstrap": "^5.3.2",
     "bootstrap-icons": "^1.11.1",
@@ -38,8 +36,6 @@
     "react-dom": "^18.2.0",
     "react-i18next": "^13.3.1",
     "react-redux": "^8.1.3",
-    "react-window": "^1.8.9",
-    "react-window-infinite-loader": "^1.0.9",
     "vite-plugin-top-level-await": "^1.3.1",
     "vscode": "npm:@codingame/monaco-vscode-api@^5.0.1"
   },
