+++ src/components/LocalStorageAutoSaver.tsx
@@ -1,9 +1,6 @@
 import { useEffect } from "react";
 import { createTimeoutEffect } from "../createTimeoutEffect";
-import {
-  programTextLocalStorageKey,
-  uiSettingLocalStorageKey,
-} from "../localStorageKeys";
+import { programTextLocalStorageKey } from "../localStorageKeys";
 import { useAppSelector } from "../store";
 import { selectProgramText } from "../store/programInfo/selectors/selectProgramText";
 
@@ -11,19 +8,14 @@ import { selectProgramText } from "../store/programInfo/selectors/selectProgramT
  * Saves the UI setting and editor state to the local storage.
  */
 export function LocalStorageAutoSaver() {
-  const uiSettings = useAppSelector((store) => store.uiSettings);
   const programText = useAppSelector(selectProgramText);
 
   // eslint-disable-next-line react-hooks/exhaustive-deps
   useEffect(
     createTimeoutEffect(() => {
       window.localStorage.setItem(programTextLocalStorageKey, programText);
-      window.localStorage.setItem(
-        uiSettingLocalStorageKey,
-        JSON.stringify(uiSettings),
-      );
     }, 500),
-    [uiSettings, programText],
+    [programText],
   );
 
   return <></>;
