+++ src/components/editor/MonacoWrapper.css
@@ -1,4 +1,4 @@
 .monaco-wrapper-container,
 .monaco-wrapper-container > div {
-  min-height: 80vh;
+  min-height: 60vh;
 }
