+++ src/components/executionPanel/results/PredicateResults.css
@@ -1,0 +1,6 @@
+.table-controls {
+  display: flex;
+  flex-flow: row wrap;
+  justify-content: space-between;
+  align-items: center;
+}
