--- src/components/leftPanel/LeftPanel.tsx
@@ -1,5 +1,0 @@
-import { Editor } from "./editor/Editor";
-
-export function LeftPanel() {
-  return <Editor />;
-}
