+++ src/components/navigationBar/NavigationBar.tsx
@@ -1,12 +1,12 @@
-import { Navbar, Container, Nav, Button } from "react-bootstrap";
+import { Navbar, Container, Nav } from "react-bootstrap";
 import "./NavigationBar.css";
-import { UISettings } from "./UISettings";
 import { useTranslation } from "react-i18next";
 import { links } from "../links";
 import { useEffect, useState } from "react";
 import { createNemoWorker } from "../../nemoWorker/NemoWorker";
-import { TextTooltip } from "../TextTooltip";
 import logoNemo from "./nemo-logo-rusty.svg";
+import { Icon } from "../Icon";
+import { ShowExamplesButton } from "./ShowExamplesButton";
 
 export function NavigationBar() {
   const { t } = useTranslation("navigationBar");
@@ -26,11 +26,11 @@ export function NavigationBar() {
   return (
     <Navbar
       variant="light"
-      expand="lg"
+      expand="md"
       style={{ borderBottom: "1px solid #dddddd" }}
     >
-      <Container>
-        <Navbar.Brand className="me-0">
+      <Container fluid>
+        <Navbar.Brand>
           <img
             src={logoNemo}
             alt="Nemo - Graph Rule Engine"
@@ -40,28 +40,21 @@ export function NavigationBar() {
         <Navbar.Toggle aria-controls="navigation-bar-nav" />
         <Navbar.Collapse id="navigation-bar-nav">
           <Nav className="me-auto">
-            <TextTooltip
-              text="Version of Nemo Rule Engine (nemo-wasm)"
-              tooltipID="navigation-bar-version-tooltip"
-            >
-              <Nav.Link className="me-2">{nemoVersion}</Nav.Link>
-            </TextTooltip>
-            <Nav.Link href={links.sourceCodeNemo} target="_blank">
-              {t("sourceCodeNemo")}
-            </Nav.Link>
-            <Nav.Link href={links.sourceCodeNemoWeb} target="_blank">
-              {t("sourceCodeNemoWeb")}
-            </Nav.Link>
             <Nav.Link href={links.documentation} target="_blank">
-              {t("documentation")}
+              {t("documentation")} <Icon name="file-earmark-code" />
             </Nav.Link>
+            <ShowExamplesButton />
           </Nav>
-          <a href={links.feedback} target="_blank">
-            <Button className="me-4" variant="primary">
-              {t("common:giveFeedback")}
-            </Button>
-          </a>
-          <UISettings />
+          <Navbar.Text>
+            <span style={{ fontSize: ".8em" }}>Version: {nemoVersion}</span>{" "}
+            <a
+              href={links.sourceCodeNemo}
+              target="_blank"
+              style={{ color: "inherit" }}
+            >
+              <Icon name="github" />
+            </a>
+          </Navbar.Text>
         </Navbar.Collapse>
       </Container>
     </Navbar>
