--- src/components/rightPanel/RightPanel.tsx
@@ -1,5 +1,0 @@
-import { ExecutionPanel } from "./executionPanel/ExecutionPanel";
-
-export function RightPanel() {
-  return <ExecutionPanel />;
-}
