--- src/components/rightPanel/executionPanel/results/PredicateResults.css
@@ -1,8 +1,0 @@
-.predicate-results-width-calculation {
-  position: absolute;
-
-  visibility: hidden;
-  height: auto;
-  width: auto;
-  white-space: nowrap;
-}
