+++ src/localStorageKeys.ts
@@ -1,2 +1,1 @@
 export const programTextLocalStorageKey = "v1-programText";
-export const uiSettingLocalStorageKey = "v1-uiSettings";
