+++ src/store/index.ts
@@ -2,12 +2,10 @@ import { configureStore } from "@reduxjs/toolkit";
 import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
 import { programInfoSlice } from "./programInfo";
 import { toastsSlice } from "./toasts";
-import { uiSettingsSlice } from "./uiSettings";
 
 export function createStore() {
   return configureStore({
     reducer: {
-      uiSettings: uiSettingsSlice.reducer,
       programInfo: programInfoSlice.reducer,
       toasts: toastsSlice.reducer,
     },
