--- src/store/uiSettings/UISettings.ts
@@ -1,5 +1,0 @@
-export interface UISettings {
-  enableFullscreen: boolean;
-  showLeftPanel: boolean;
-  showRightPanel: boolean;
-}
