--- src/store/uiSettings/initialUISettings.ts
@@ -1,7 +1,0 @@
-import { UISettings } from "./UISettings";
-
-export const initialUISettings: UISettings = {
-  enableFullscreen: false,
-  showLeftPanel: true,
-  showRightPanel: false,
-};
