--- src/store/uiSettings/reducers/setAllUISettings.ts
@@ -1,12 +1,0 @@
-import { PayloadAction } from "@reduxjs/toolkit";
-import { UISettings } from "../UISettings";
-
-export function setAllUISettings(
-  state: UISettings,
-  action: PayloadAction<UISettings>,
-): UISettings {
-  return {
-    ...state,
-    ...action.payload,
-  };
-}
