--- src/store/uiSettings/reducers/toggleFullscreen.ts
@@ -1,12 +1,0 @@
-import { PayloadAction } from "@reduxjs/toolkit";
-import { UISettings } from "../UISettings";
-
-export function toggleFullscreen(
-  state: UISettings,
-  _action: PayloadAction<void>,
-): UISettings {
-  return {
-    ...state,
-    enableFullscreen: !state.enableFullscreen,
-  };
-}
